+++ CSC3224 - Game Development Project/CSC3224 - Graphics/Circle.h
@@ -3,16 +3,15 @@
 #define CIRCLE_H
 
 #include <SFML\Graphics.hpp>
-
 using namespace std;
 
-class Circle 
+class Circle
 {
 public:
 	Circle(); //Default constructor
 	virtual ~Circle() {}; //Destructor
 
-	void createCircle(const string &name, const int &radius); //Function to create an rectangle
+	void createCircle(const string &name, const float &radius); //Function to create an rectangle
 	void clipCircleTex(const sf::IntRect &clip); //Function to clip a texture applied to a circle
 	sf::CircleShape &getCircleRef(const string &circle); //Function to fetch the reference to the specified circle
